+++ pybutton/constants.py
@@ -1,0 +1,2 @@
+TIME_FIELD_CREATED = 'created_date'
+TIME_FIELD_MODIFIED = 'modified_date'